[-PDE] Replace all .pde files by the .java product
SPaTo has outgrown the typical Processing sketch size a while ago
already.  Meanwhile, we even post-process the applications generated by
the Processing Development Environment (PDE) for all three platforms, so
at this point the PDE is merely used for compilation and poses an
unnecessary barrier for other people to build the application from the
sources.

Thus, I will reorganize the complete codebase so that it can be compiled
by the Java compiler directly and set up an ant build system.  SPaTo
will continue to use the Processing core.jar, but won't require the PDE
anymore.  Updating from the custom core.jar currently used will be the
next major refactoring step, once Processing 2.0 is released.

All commits related to this emancipation process will be tagged [-PDE]
and might contain broken builds of the application (that can't be
compiled), such as this one...