Simplify SPaTo Prelude
The prelude was run in its own process because it would measure
the maximum memory that we can allocate for Java and then launch
the application in a new JVM.  Since we are not doing maximum
memory detection anymore, the separation into two processes is
unnecessary and only increases application start-up time.

Thus, the prelude has been completely reworked so that both the
prelude and the application run within the same JVM.  Due to a
bug in the self-update mechanism of SPaTo_Prelude we have to
provide UpdateInstaller in the default package or otherwise the
user would have to manually update beyond version 1.2.2.

The prelude as in this commit is tested on Mac OS X and will
successfully auto-update from version 1.2.2.  Testing on
Windows and Linux is yet to be done...